Split s3
