+++ frictionless/plugins/s3/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import S3Plugin
+from .control import S3Control
+from .loader import S3Loader
